move all DBC-specific code into a separate sub-module
Approved-by: Alexander Walz <alexander.walz@mercedes-benz.com>
