+++ docs/index.rst
@@ -40,19 +40,19 @@ Functions and classes
 .. autoclass:: cantools.database.can.signal_group.SignalGroup
     :members:
 
-.. autoclass:: cantools.database.can.attribute_definition.AttributeDefinition
+.. autoclass:: cantools.database.can.formats.dbc.attribute_definition.AttributeDefinition
     :members:
 
-.. autoclass:: cantools.database.can.attribute_definition.AttributeValueType
+.. autoclass:: cantools.database.can.formats.dbc.attribute_definition.AttributeValueType
     :members:
 
-.. autoclass:: cantools.database.can.attribute.Attribute
+.. autoclass:: cantools.database.can.formats.dbc.attribute.Attribute
     :members:
 
-.. autoclass:: cantools.database.can.environment_variable.EnvironmentVariable
+.. autoclass:: cantools.database.can.formats.dbc.environment_variable.EnvironmentVariable
     :members:
 
-.. autoclass:: cantools.database.can.formats.dbc_specifics.DbcSpecifics
+.. autoclass:: cantools.database.can.formats.dbc.dbc_specifics.DbcSpecifics
     :members:
 
 .. autoclass:: cantools.database.can.formats.arxml.AutosarDatabaseSpecifics
