+++ src/cantools/database/can/formats/dbc/__init__.py
@@ -1,0 +1,21 @@
+from .attribute import Attribute, AttributeType
+from .attribute_definition import (
+    AttributeDefinition,
+    AttributeDefinitionType,
+    AttributeValueType,
+)
+from .dbc import dump_string, load_string
+from .dbc_specifics import DbcSpecifics
+from .environment_variable import EnvironmentVariable
+
+__all__ = [
+    "Attribute",
+    "AttributeDefinition",
+    "AttributeDefinitionType",
+    "AttributeType",
+    "AttributeValueType",
+    "DbcSpecifics",
+    "EnvironmentVariable",
+    "dump_string",
+    "load_string",
+]
