+++ src/cantools/database/can/formats/dbc/attribute.py
