+++ src/cantools/database/can/formats/dbc/attribute_definition.py
